fix(*): only freeze when supported
React Native on Android used a fucking retarded JavaScript engine that cannot create a `Set` of frozen objects. I assume it is because it polyfills the `Set` class and that adds some hidden variable to the object that cannot be added because the object is frozen. Maybe. Or it might be because whatever engine they use it bloody ancient and they love inflicting pain on the whole React Native community debugging this fucking bullshit